requirements.txt
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-JWT-Extended==3.24.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==1.6.4
pymongo==3.11.2
python-dotenv==0.8.2
six==1.15.0
Werkzeug==1.0.1
